// src/app/shared/modules/alerts/enums/alerts.enum.ts
export enum AlertTypes {
    Success = 'success',
    Error = 'error',
    Alert = 'alert',
    Info = 'info',
    Warn = 'warn',
    Bare = 'bare',
}

export enum AnimationType {
    Fade = 'fade',
    FromTop = 'fromTop',
    FromRight = 'fromRight',
    FromBottom = 'fromBottom',
    FromLeft = 'fromLeft',
    Scale = 'scale',
    Rotate = 'rotate',
}